refactor: address PR feedback
