--- src/app/profile/certifications/components/certificationList/index.ts
@@ -1,5 +1,0 @@
-import CertificationList from './certificationList';
-
-export {
-  CertificationList,
-}
