--- src/app/profile/certifications/components/certificationUploadForm/index.ts
@@ -1,5 +1,0 @@
-import CertificationUploadForm from './certificationUploadForm';
-
-export {
-  CertificationUploadForm,
-}
