--- src/app/profile/components/navigationTabs/index.ts
@@ -1,5 +1,0 @@
-import NavigationTabs from './NavigationTabs';
-
-export {
-  NavigationTabs,
-}
