--- src/app/profile/components/organisationProfileForm/index.ts
@@ -1,5 +1,0 @@
-import OrganisationProfileForm from './organisationProfileForm';
-
-export {
-  OrganisationProfileForm,
-}
