--- src/app/profile/components/userProfileForm/index.ts
@@ -1,5 +1,0 @@
-import UserProfileForm from './userProfileForm';
-
-export {
-  UserProfileForm,
-}
