--- src/shared/button/index.ts
@@ -1,5 +1,0 @@
-import Button from './Button';
-
-export {
-  Button,
-}
