--- src/shared/form/index.ts
@@ -1,6 +1,0 @@
-import Form, { FieldError } from './Form';
-
-export {
-  Form,
-  FieldError,
-}
