--- src/shared/formInput/index.ts
@@ -1,5 +1,0 @@
-import FormInput from './FormInput';
-
-export {
-  FormInput,
-}
