--- src/shared/formSelect/index.ts
@@ -1,5 +1,0 @@
-import FormSelect from './FormSelect';
-
-export {
-  FormSelect,
-}
