--- src/shared/loadingSpinner/index.ts
@@ -1,5 +1,0 @@
-import LoadingSpinner from './LoadingSpinner';
-
-export {
-  LoadingSpinner,
-}
